Fix the precompiled vignette not building.
It was not able to find the figures generated by the precompilation.
This is because it was given fig.path = vignettes/..., which caused
them to be saved in the correct place, but to be unable to find them
when building, as vignettes are built with the vignette directory
already as the working directory.
Before aed304652cf83c198a1080a19014df28d05908b0, the figure paths
were correct, though how it generated the correct figure paths is
beyond me. When I recompiled the vignettes for that commit, it
made the figure paths wrong.