Fix a bug in df_mort().
This was a copy-paste error from df_funcresp(). The data frame in
df_funcresp() has resources as the first column, so the line to
get the parameter values has to drop the resources column.
In df_mort(), tht column does not exist, so dropping the first
column is a bug.